feat(plugin): implement optimistic concurrency control plugin
Remove example module and tests.

Add updateIfCurrentPlugin function as a mongoose plugin. Set schema
to throw an error is a save operation fails due to the document not
being found. Enable the default version key on the schema if one is
not already set. Use a pre-save hook to add the current version as
a $where condition for the save operation and to increment the
version number.

Add test cases to cover the use of the plugin.